requirements.txt
groq==0.4.2
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
openai==1.3.7
langchain==0.0.350
langchain-groq==0.0.1
streamlit==1.28.2
plotly==5.17.0
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
